transmute for hyperSpec objects

adds new variables and drops existing ones. Special column $spc contains the spectra matrix. Unlike dplyr::transmute() dplyr::mutate() preserves all the existing variables. If $spc is not being transmuted, the result is a data.frame instead of a hyperSpec object.

# S3 method for hyperSpec
transmute(.data, ...)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<data-masking> Name-value pairs. The name gives the name of the column in the output.

The value can be:

A vector of length 1, which will be recycled to the correct length.

A vector the same length as the current group (or the whole data frame if ungrouped).

NULL, to remove the column.

A data frame or tibble, to create multiple columns in the output.

Value

hyperSpec object. If $spc is renamed, the result is an error.

See also

Examples

chondro %>%
  transmute (y, x, filename, spc) %>%
  head # => results in a hyperSpec object

#> hyperSpec object
#>    6 spectra
#>    4 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (6 rows x 4 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... -4.77 
#>    2. x: x [numeric] -11.55 -10.55 ... -6.55 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 184.0021 

chondro %>%
  transmute (y, x)

#>         y      x
#> 1   -4.77 -11.55
#> 2   -4.77 -10.55
#> 3   -4.77  -9.55
#> 4   -4.77  -8.55
#> 5   -4.77  -7.55
#> 6   -4.77  -6.55
#> 7   -4.77  -5.55
#> 8   -4.77  -4.55
#> 9   -4.77  -3.55
#> 10  -4.77  -2.55
#> 11  -4.77  -1.55
#> 12  -4.77  -0.55
#> 13  -4.77   0.45
#> 14  -4.77   1.45
#> 15  -4.77   2.45
#> 16  -4.77   3.45
#> 17  -4.77   4.45
#> 18  -4.77   5.45
#> 19  -4.77   6.45
#> 20  -4.77   7.45
#> 21  -4.77   8.45
#> 22  -4.77   9.45
#> 23  -4.77  10.45
#> 24  -4.77  11.45
#> 25  -4.77  12.45
#> 26  -4.77  13.45
#> 27  -4.77  14.45
#> 28  -4.77  15.45
#> 29  -4.77  16.45
#> 30  -4.77  17.45
#> 31  -4.77  18.45
#> 32  -4.77  19.45
#> 33  -4.77  20.45
#> 34  -4.77  21.45
#> 35  -4.77  22.45
#> 36  -3.77 -11.55
#> 37  -3.77 -10.55
#> 38  -3.77  -9.55
#> 39  -3.77  -8.55
#> 40  -3.77  -7.55
#> 41  -3.77  -6.55
#> 42  -3.77  -5.55
#> 43  -3.77  -4.55
#> 44  -3.77  -3.55
#> 45  -3.77  -2.55
#> 46  -3.77  -1.55
#> 47  -3.77  -0.55
#> 48  -3.77   0.45
#> 49  -3.77   1.45
#> 50  -3.77   2.45
#> 51  -3.77   3.45
#> 52  -3.77   4.45
#> 53  -3.77   5.45
#> 54  -3.77   6.45
#> 55  -3.77   7.45
#> 56  -3.77   8.45
#> 57  -3.77   9.45
#> 58  -3.77  10.45
#> 59  -3.77  11.45
#> 60  -3.77  12.45
#> 61  -3.77  13.45
#> 62  -3.77  14.45
#> 63  -3.77  15.45
#> 64  -3.77  16.45
#> 65  -3.77  17.45
#> 66  -3.77  18.45
#> 67  -3.77  19.45
#> 68  -3.77  20.45
#> 69  -3.77  21.45
#> 70  -3.77  22.45
#> 71  -2.77 -11.55
#> 72  -2.77 -10.55
#> 73  -2.77  -9.55
#> 74  -2.77  -8.55
#> 75  -2.77  -7.55
#> 76  -2.77  -6.55
#> 77  -2.77  -5.55
#> 78  -2.77  -4.55
#> 79  -2.77  -3.55
#> 80  -2.77  -2.55
#> 81  -2.77  -1.55
#> 82  -2.77  -0.55
#> 83  -2.77   0.45
#> 84  -2.77   1.45
#> 85  -2.77   2.45
#> 86  -2.77   3.45
#> 87  -2.77   4.45
#> 88  -2.77   5.45
#> 89  -2.77   6.45
#> 90  -2.77   7.45
#> 91  -2.77   8.45
#> 92  -2.77   9.45
#> 93  -2.77  10.45
#> 94  -2.77  11.45
#> 95  -2.77  12.45
#> 96  -2.77  13.45
#> 97  -2.77  14.45
#> 98  -2.77  15.45
#> 99  -2.77  16.45
#> 100 -2.77  17.45
#> 101 -2.77  18.45
#> 102 -2.77  19.45
#> 103 -2.77  20.45
#> 104 -2.77  21.45
#> 105 -2.77  22.45
#> 106 -1.77 -11.55
#> 107 -1.77 -10.55
#> 108 -1.77  -9.55
#> 109 -1.77  -8.55
#> 110 -1.77  -7.55
#> 111 -1.77  -6.55
#> 112 -1.77  -5.55
#> 113 -1.77  -4.55
#> 114 -1.77  -3.55
#> 115 -1.77  -2.55
#> 116 -1.77  -1.55
#> 117 -1.77  -0.55
#> 118 -1.77   0.45
#> 119 -1.77   1.45
#> 120 -1.77   2.45
#> 121 -1.77   3.45
#> 122 -1.77   4.45
#> 123 -1.77   5.45
#> 124 -1.77   6.45
#> 125 -1.77   7.45
#> 126 -1.77   8.45
#> 127 -1.77   9.45
#> 128 -1.77  10.45
#> 129 -1.77  11.45
#> 130 -1.77  12.45
#> 131 -1.77  13.45
#> 132 -1.77  14.45
#> 133 -1.77  15.45
#> 134 -1.77  16.45
#> 135 -1.77  17.45
#> 136 -1.77  18.45
#> 137 -1.77  19.45
#> 138 -1.77  20.45
#> 139 -1.77  21.45
#> 140 -1.77  22.45
#> 141 -0.77 -11.55
#> 142 -0.77 -10.55
#> 143 -0.77  -9.55
#> 144 -0.77  -8.55
#> 145 -0.77  -7.55
#> 146 -0.77  -6.55
#> 147 -0.77  -5.55
#> 148 -0.77  -4.55
#> 149 -0.77  -3.55
#> 150 -0.77  -2.55
#> 151 -0.77  -1.55
#> 152 -0.77  -0.55
#> 153 -0.77   0.45
#> 154 -0.77   1.45
#> 155 -0.77   2.45
#> 156 -0.77   3.45
#> 157 -0.77   4.45
#> 158 -0.77   5.45
#> 159 -0.77   6.45
#> 160 -0.77   7.45
#> 161 -0.77   8.45
#> 162 -0.77   9.45
#> 163 -0.77  10.45
#> 164 -0.77  11.45
#> 165 -0.77  12.45
#> 166 -0.77  13.45
#> 167 -0.77  14.45
#> 168 -0.77  15.45
#> 169 -0.77  16.45
#> 170 -0.77  17.45
#> 171 -0.77  18.45
#> 172 -0.77  19.45
#> 173 -0.77  20.45
#> 174 -0.77  21.45
#> 175 -0.77  22.45
#> 176  0.23 -11.55
#> 177  0.23 -10.55
#> 178  0.23  -9.55
#> 179  0.23  -8.55
#> 180  0.23  -7.55
#> 181  0.23  -6.55
#> 182  0.23  -5.55
#> 183  0.23  -4.55
#> 184  0.23  -3.55
#> 185  0.23  -2.55
#> 186  0.23  -1.55
#> 187  0.23  -0.55
#> 188  0.23   0.45
#> 189  0.23   1.45
#> 190  0.23   2.45
#> 191  0.23   3.45
#> 192  0.23   4.45
#> 193  0.23   5.45
#> 194  0.23   6.45
#> 195  0.23   7.45
#> 196  0.23   8.45
#> 197  0.23   9.45
#> 198  0.23  10.45
#> 199  0.23  11.45
#> 200  0.23  12.45
#> 201  0.23  13.45
#> 202  0.23  14.45
#> 203  0.23  15.45
#> 204  0.23  16.45
#> 205  0.23  17.45
#> 206  0.23  18.45
#> 207  0.23  19.45
#> 208  0.23  20.45
#> 209  0.23  21.45
#> 210  0.23  22.45
#> 211  1.23 -11.55
#> 212  1.23 -10.55
#> 213  1.23  -9.55
#> 214  1.23  -8.55
#> 215  1.23  -7.55
#> 216  1.23  -6.55
#> 217  1.23  -5.55
#> 218  1.23  -4.55
#> 219  1.23  -3.55
#> 220  1.23  -2.55
#> 221  1.23  -1.55
#> 222  1.23  -0.55
#> 223  1.23   0.45
#> 224  1.23   1.45
#> 225  1.23   2.45
#> 226  1.23   3.45
#> 227  1.23   4.45
#> 228  1.23   5.45
#> 229  1.23   6.45
#> 230  1.23   7.45
#> 231  1.23   8.45
#> 232  1.23   9.45
#> 233  1.23  10.45
#> 234  1.23  11.45
#> 235  1.23  12.45
#> 236  1.23  13.45
#> 237  1.23  14.45
#> 238  1.23  15.45
#> 239  1.23  16.45
#> 240  1.23  17.45
#> 241  1.23  18.45
#> 242  1.23  19.45
#> 243  1.23  20.45
#> 244  1.23  21.45
#> 245  1.23  22.45
#> 246  2.23 -11.55
#> 247  2.23 -10.55
#> 248  2.23  -9.55
#> 249  2.23  -8.55
#> 250  2.23  -7.55
#> 251  2.23  -6.55
#> 252  2.23  -5.55
#> 253  2.23  -4.55
#> 254  2.23  -3.55
#> 255  2.23  -2.55
#> 256  2.23  -1.55
#> 257  2.23  -0.55
#> 258  2.23   0.45
#> 259  2.23   1.45
#> 260  2.23   2.45
#> 261  2.23   3.45
#> 262  2.23   4.45
#> 263  2.23   5.45
#> 264  2.23   6.45
#> 265  2.23   7.45
#> 266  2.23   8.45
#> 267  2.23   9.45
#> 268  2.23  10.45
#> 269  2.23  11.45
#> 270  2.23  12.45
#> 271  2.23  13.45
#> 272  2.23  14.45
#> 273  2.23  15.45
#> 274  2.23  16.45
#> 275  2.23  17.45
#> 276  2.23  18.45
#> 277  2.23  19.45
#> 278  2.23  20.45
#> 279  2.23  21.45
#> 280  2.23  22.45
#> 281  3.23 -11.55
#> 282  3.23 -10.55
#> 283  3.23  -9.55
#> 284  3.23  -8.55
#> 285  3.23  -7.55
#> 286  3.23  -6.55
#> 287  3.23  -5.55
#> 288  3.23  -4.55
#> 289  3.23  -3.55
#> 290  3.23  -2.55
#> 291  3.23  -1.55
#> 292  3.23  -0.55
#> 293  3.23   0.45
#> 294  3.23   1.45
#> 295  3.23   2.45
#> 296  3.23   3.45
#> 297  3.23   4.45
#> 298  3.23   5.45
#> 299  3.23   6.45
#> 300  3.23   7.45
#> 301  3.23   8.45
#> 302  3.23   9.45
#> 303  3.23  10.45
#> 304  3.23  11.45
#> 305  3.23  12.45
#> 306  3.23  13.45
#> 307  3.23  14.45
#> 308  3.23  15.45
#> 309  3.23  16.45
#> 310  3.23  17.45
#> 311  3.23  18.45
#> 312  3.23  19.45
#> 313  3.23  20.45
#> 314  3.23  21.45
#> 315  3.23  22.45
#> 316  4.23 -11.55
#> 317  4.23 -10.55
#> 318  4.23  -9.55
#> 319  4.23  -8.55
#> 320  4.23  -7.55
#> 321  4.23  -6.55
#> 322  4.23  -5.55
#> 323  4.23  -4.55
#> 324  4.23  -3.55
#> 325  4.23  -2.55
#> 326  4.23  -1.55
#> 327  4.23  -0.55
#> 328  4.23   0.45
#> 329  4.23   1.45
#> 330  4.23   2.45
#> 331  4.23   3.45
#> 332  4.23   4.45
#> 333  4.23   5.45
#> 334  4.23   6.45
#> 335  4.23   7.45
#> 336  4.23   8.45
#> 337  4.23   9.45
#> 338  4.23  10.45
#> 339  4.23  11.45
#> 340  4.23  12.45
#> 341  4.23  13.45
#> 342  4.23  14.45
#> 343  4.23  15.45
#> 344  4.23  16.45
#> 345  4.23  17.45
#> 346  4.23  18.45
#> 347  4.23  19.45
#> 348  4.23  20.45
#> 349  4.23  21.45
#> 350  4.23  22.45
#> 351  5.23 -11.55
#> 352  5.23 -10.55
#> 353  5.23  -9.55
#> 354  5.23  -8.55
#> 355  5.23  -7.55
#> 356  5.23  -6.55
#> 357  5.23  -5.55
#> 358  5.23  -4.55
#> 359  5.23  -3.55
#> 360  5.23  -2.55
#> 361  5.23  -1.55
#> 362  5.23  -0.55
#> 363  5.23   0.45
#> 364  5.23   1.45
#> 365  5.23   2.45
#> 366  5.23   3.45
#> 367  5.23   4.45
#> 368  5.23   5.45
#> 369  5.23   6.45
#> 370  5.23   7.45
#> 371  5.23   8.45
#> 372  5.23   9.45
#> 373  5.23  10.45
#> 374  5.23  11.45
#> 375  5.23  12.45
#> 376  5.23  13.45
#> 377  5.23  14.45
#> 378  5.23  15.45
#> 379  5.23  16.45
#> 380  5.23  17.45
#> 381  5.23  18.45
#> 382  5.23  19.45
#> 383  5.23  20.45
#> 384  5.23  21.45
#> 385  5.23  22.45
#> 386  6.23 -11.55
#> 387  6.23 -10.55
#> 388  6.23  -9.55
#> 389  6.23  -8.55
#> 390  6.23  -7.55
#> 391  6.23  -6.55
#> 392  6.23  -5.55
#> 393  6.23  -4.55
#> 394  6.23  -3.55
#> 395  6.23  -2.55
#> 396  6.23  -1.55
#> 397  6.23  -0.55
#> 398  6.23   0.45
#> 399  6.23   1.45
#> 400  6.23   2.45
#> 401  6.23   3.45
#> 402  6.23   4.45
#> 403  6.23   5.45
#> 404  6.23   6.45
#> 405  6.23   7.45
#> 406  6.23   8.45
#> 407  6.23   9.45
#> 408  6.23  10.45
#> 409  6.23  11.45
#> 410  6.23  12.45
#> 411  6.23  13.45
#> 412  6.23  14.45
#> 413  6.23  15.45
#> 414  6.23  16.45
#> 415  6.23  17.45
#> 416  6.23  18.45
#> 417  6.23  19.45
#> 418  6.23  20.45
#> 419  6.23  21.45
#> 420  6.23  22.45
#> 421  7.23 -11.55
#> 422  7.23 -10.55
#> 423  7.23  -9.55
#> 424  7.23  -8.55
#> 425  7.23  -7.55
#> 426  7.23  -6.55
#> 427  7.23  -5.55
#> 428  7.23  -4.55
#> 429  7.23  -3.55
#> 430  7.23  -2.55
#> 431  7.23  -1.55
#> 432  7.23  -0.55
#> 433  7.23   0.45
#> 434  7.23   1.45
#> 435  7.23   2.45
#> 436  7.23   3.45
#> 437  7.23   4.45
#> 438  7.23   5.45
#> 439  7.23   6.45
#> 440  7.23   7.45
#> 441  7.23   8.45
#> 442  7.23   9.45
#> 443  7.23  10.45
#> 444  7.23  11.45
#> 445  7.23  12.45
#> 446  7.23  13.45
#> 447  7.23  14.45
#> 448  7.23  15.45
#> 449  7.23  16.45
#> 450  7.23  17.45
#> 451  7.23  18.45
#> 452  7.23  19.45
#> 453  7.23  20.45
#> 454  7.23  21.45
#> 455  7.23  22.45
#> 456  8.23 -11.55
#> 457  8.23 -10.55
#> 458  8.23  -9.55
#> 459  8.23  -8.55
#> 460  8.23  -7.55
#> 461  8.23  -6.55
#> 462  8.23  -5.55
#> 463  8.23  -4.55
#> 464  8.23  -3.55
#> 465  8.23  -2.55
#> 466  8.23  -1.55
#> 467  8.23  -0.55
#> 468  8.23   0.45
#> 469  8.23   1.45
#> 470  8.23   2.45
#> 471  8.23   3.45
#> 472  8.23   4.45
#> 473  8.23   5.45
#> 474  8.23   6.45
#> 475  8.23   7.45
#> 476  8.23   8.45
#> 477  8.23   9.45
#> 478  8.23  10.45
#> 479  8.23  11.45
#> 480  8.23  12.45
#> 481  8.23  13.45
#> 482  8.23  14.45
#> 483  8.23  15.45
#> 484  8.23  16.45
#> 485  8.23  17.45
#> 486  8.23  18.45
#> 487  8.23  19.45
#> 488  8.23  20.45
#> 489  8.23  21.45
#> 490  8.23  22.45
#> 491  9.23 -11.55
#> 492  9.23 -10.55
#> 493  9.23  -9.55
#> 494  9.23  -8.55
#> 495  9.23  -7.55
#> 496  9.23  -6.55
#> 497  9.23  -5.55
#> 498  9.23  -4.55
#> 499  9.23  -3.55
#> 500  9.23  -2.55
#> 501  9.23  -1.55
#> 502  9.23  -0.55
#> 503  9.23   0.45
#> 504  9.23   1.45
#> 505  9.23   2.45
#> 506  9.23   3.45
#> 507  9.23   4.45
#> 508  9.23   5.45
#> 509  9.23   6.45
#> 510  9.23   7.45
#> 511  9.23   8.45
#> 512  9.23   9.45
#> 513  9.23  10.45
#> 514  9.23  11.45
#> 515  9.23  12.45
#> 516  9.23  13.45
#> 517  9.23  14.45
#> 518  9.23  15.45
#> 519  9.23  16.45
#> 520  9.23  17.45
#> 521  9.23  18.45
#> 522  9.23  19.45
#> 523  9.23  20.45
#> 524  9.23  21.45
#> 525  9.23  22.45
#> 526 10.23 -11.55
#> 527 10.23 -10.55
#> 528 10.23  -9.55
#> 529 10.23  -8.55
#> 530 10.23  -7.55
#> 531 10.23  -6.55
#> 532 10.23  -5.55
#> 533 10.23  -4.55
#> 534 10.23  -3.55
#> 535 10.23  -2.55
#> 536 10.23  -1.55
#> 537 10.23  -0.55
#> 538 10.23   0.45
#> 539 10.23   1.45
#> 540 10.23   2.45
#> 541 10.23   3.45
#> 542 10.23   4.45
#> 543 10.23   5.45
#> 544 10.23   6.45
#> 545 10.23   7.45
#> 546 10.23   8.45
#> 547 10.23   9.45
#> 548 10.23  10.45
#> 549 10.23  11.45
#> 550 10.23  12.45
#> 551 10.23  13.45
#> 552 10.23  14.45
#> 553 10.23  15.45
#> 554 10.23  16.45
#> 555 10.23  17.45
#> 556 10.23  18.45
#> 557 10.23  19.45
#> 558 10.23  20.45
#> 559 10.23  21.45
#> 560 10.23  22.45
#> 561 11.23 -11.55
#> 562 11.23 -10.55
#> 563 11.23  -9.55
#> 564 11.23  -8.55
#> 565 11.23  -7.55
#> 566 11.23  -6.55
#> 567 11.23  -5.55
#> 568 11.23  -4.55
#> 569 11.23  -3.55
#> 570 11.23  -2.55
#> 571 11.23  -1.55
#> 572 11.23  -0.55
#> 573 11.23   0.45
#> 574 11.23   1.45
#> 575 11.23   2.45
#> 576 11.23   3.45
#> 577 11.23   4.45
#> 578 11.23   5.45
#> 579 11.23   6.45
#> 580 11.23   7.45
#> 581 11.23   8.45
#> 582 11.23   9.45
#> 583 11.23  10.45
#> 584 11.23  11.45
#> 585 11.23  12.45
#> 586 11.23  13.45
#> 587 11.23  14.45
#> 588 11.23  15.45
#> 589 11.23  16.45
#> 590 11.23  17.45
#> 591 11.23  18.45
#> 592 11.23  19.45
#> 593 11.23  20.45
#> 594 11.23  21.45
#> 595 11.23  22.45
#> 596 12.23 -11.55
#> 597 12.23 -10.55
#> 598 12.23  -9.55
#> 599 12.23  -8.55
#> 600 12.23  -7.55
#> 601 12.23  -6.55
#> 602 12.23  -5.55
#> 603 12.23  -4.55
#> 604 12.23  -3.55
#> 605 12.23  -2.55
#> 606 12.23  -1.55
#> 607 12.23  -0.55
#> 608 12.23   0.45
#> 609 12.23   1.45
#> 610 12.23   2.45
#> 611 12.23   3.45
#> 612 12.23   4.45
#> 613 12.23   5.45
#> 614 12.23   6.45
#> 615 12.23   7.45
#> 616 12.23   8.45
#> 617 12.23   9.45
#> 618 12.23  10.45
#> 619 12.23  11.45
#> 620 12.23  12.45
#> 621 12.23  13.45
#> 622 12.23  14.45
#> 623 12.23  15.45
#> 624 12.23  16.45
#> 625 12.23  17.45
#> 626 12.23  18.45
#> 627 12.23  19.45
#> 628 12.23  20.45
#> 629 12.23  21.45
#> 630 12.23  22.45
#> 631 13.23 -11.55
#> 632 13.23 -10.55
#> 633 13.23  -9.55
#> 634 13.23  -8.55
#> 635 13.23  -7.55
#> 636 13.23  -6.55
#> 637 13.23  -5.55
#> 638 13.23  -4.55
#> 639 13.23  -3.55
#> 640 13.23  -2.55
#> 641 13.23  -1.55
#> 642 13.23  -0.55
#> 643 13.23   0.45
#> 644 13.23   1.45
#> 645 13.23   2.45
#> 646 13.23   3.45
#> 647 13.23   4.45
#> 648 13.23   5.45
#> 649 13.23   6.45
#> 650 13.23   7.45
#> 651 13.23   8.45
#> 652 13.23   9.45
#> 653 13.23  10.45
#> 654 13.23  11.45
#> 655 13.23  12.45
#> 656 13.23  13.45
#> 657 13.23  14.45
#> 658 13.23  15.45
#> 659 13.23  16.45
#> 660 13.23  17.45
#> 661 13.23  18.45
#> 662 13.23  19.45
#> 663 13.23  20.45
#> 664 13.23  21.45
#> 665 13.23  22.45
#> 666 14.23 -11.55
#> 667 14.23 -10.55
#> 668 14.23  -9.55
#> 669 14.23  -8.55
#> 670 14.23  -7.55
#> 671 14.23  -6.55
#> 672 14.23  -5.55
#> 673 14.23  -4.55
#> 674 14.23  -3.55
#> 675 14.23  -2.55
#> 676 14.23  -1.55
#> 677 14.23  -0.55
#> 678 14.23   0.45
#> 679 14.23   1.45
#> 680 14.23   2.45
#> 681 14.23   3.45
#> 682 14.23   4.45
#> 683 14.23   5.45
#> 684 14.23   6.45
#> 685 14.23   7.45
#> 686 14.23   8.45
#> 687 14.23   9.45
#> 688 14.23  10.45
#> 689 14.23  11.45
#> 690 14.23  12.45
#> 691 14.23  13.45
#> 692 14.23  14.45
#> 693 14.23  15.45
#> 694 14.23  16.45
#> 695 14.23  17.45
#> 696 14.23  18.45
#> 697 14.23  19.45
#> 698 14.23  20.45
#> 699 14.23  21.45
#> 700 14.23  22.45
#> 701 15.23 -11.55
#> 702 15.23 -10.55
#> 703 15.23  -9.55
#> 704 15.23  -8.55
#> 705 15.23  -7.55
#> 706 15.23  -6.55
#> 707 15.23  -5.55
#> 708 15.23  -4.55
#> 709 15.23  -3.55
#> 710 15.23  -2.55
#> 711 15.23  -1.55
#> 712 15.23  -0.55
#> 713 15.23   0.45
#> 714 15.23   1.45
#> 715 15.23   2.45
#> 716 15.23   3.45
#> 717 15.23   4.45
#> 718 15.23   5.45
#> 719 15.23   6.45
#> 720 15.23   7.45
#> 721 15.23   8.45
#> 722 15.23   9.45
#> 723 15.23  10.45
#> 724 15.23  11.45
#> 725 15.23  12.45
#> 726 15.23  13.45
#> 727 15.23  14.45
#> 728 15.23  15.45
#> 729 15.23  16.45
#> 730 15.23  17.45
#> 731 15.23  18.45
#> 732 15.23  19.45
#> 733 15.23  20.45
#> 734 15.23  21.45
#> 735 15.23  22.45
#> 736 16.23 -11.55
#> 737 16.23 -10.55
#> 738 16.23  -9.55
#> 739 16.23  -8.55
#> 740 16.23  -7.55
#> 741 16.23  -6.55
#> 742 16.23  -5.55
#> 743 16.23  -4.55
#> 744 16.23  -3.55
#> 745 16.23  -2.55
#> 746 16.23  -1.55
#> 747 16.23  -0.55
#> 748 16.23   0.45
#> 749 16.23   1.45
#> 750 16.23   2.45
#> 751 16.23   3.45
#> 752 16.23   4.45
#> 753 16.23   5.45
#> 754 16.23   6.45
#> 755 16.23   7.45
#> 756 16.23   8.45
#> 757 16.23   9.45
#> 758 16.23  10.45
#> 759 16.23  11.45
#> 760 16.23  12.45
#> 761 16.23  13.45
#> 762 16.23  14.45
#> 763 16.23  15.45
#> 764 16.23  16.45
#> 765 16.23  17.45
#> 766 16.23  18.45
#> 767 16.23  19.45
#> 768 16.23  20.45
#> 769 16.23  21.45
#> 770 16.23  22.45
#> 771 17.23 -11.55
#> 772 17.23 -10.55
#> 773 17.23  -9.55
#> 774 17.23  -8.55
#> 775 17.23  -7.55
#> 776 17.23  -6.55
#> 777 17.23  -5.55
#> 778 17.23  -4.55
#> 779 17.23  -3.55
#> 780 17.23  -2.55
#> 781 17.23  -1.55
#> 782 17.23  -0.55
#> 783 17.23   0.45
#> 784 17.23   1.45
#> 785 17.23   2.45
#> 786 17.23   3.45
#> 787 17.23   4.45
#> 788 17.23   5.45
#> 789 17.23   6.45
#> 790 17.23   7.45
#> 791 17.23   8.45
#> 792 17.23   9.45
#> 793 17.23  10.45
#> 794 17.23  11.45
#> 795 17.23  12.45
#> 796 17.23  13.45
#> 797 17.23  14.45
#> 798 17.23  15.45
#> 799 17.23  16.45
#> 800 17.23  17.45
#> 801 17.23  18.45
#> 802 17.23  19.45
#> 803 17.23  20.45
#> 804 17.23  21.45
#> 805 17.23  22.45
#> 806 18.23 -11.55
#> 807 18.23 -10.55
#> 808 18.23  -9.55
#> 809 18.23  -8.55
#> 810 18.23  -7.55
#> 811 18.23  -6.55
#> 812 18.23  -5.55
#> 813 18.23  -4.55
#> 814 18.23  -3.55
#> 815 18.23  -2.55
#> 816 18.23  -1.55
#> 817 18.23  -0.55
#> 818 18.23   0.45
#> 819 18.23   1.45
#> 820 18.23   2.45
#> 821 18.23   3.45
#> 822 18.23   4.45
#> 823 18.23   5.45
#> 824 18.23   6.45
#> 825 18.23   7.45
#> 826 18.23   8.45
#> 827 18.23   9.45
#> 828 18.23  10.45
#> 829 18.23  11.45
#> 830 18.23  12.45
#> 831 18.23  13.45
#> 832 18.23  14.45
#> 833 18.23  15.45
#> 834 18.23  16.45
#> 835 18.23  17.45
#> 836 18.23  18.45
#> 837 18.23  19.45
#> 838 18.23  20.45
#> 839 18.23  21.45
#> 840 18.23  22.45
#> 841 19.23 -11.55
#> 842 19.23 -10.55
#> 843 19.23  -9.55
#> 844 19.23  -8.55
#> 845 19.23  -7.55
#> 846 19.23  -6.55
#> 847 19.23  -5.55
#> 848 19.23  -4.55
#> 849 19.23  -3.55
#> 850 19.23  -2.55
#> 851 19.23  -1.55
#> 852 19.23  -0.55
#> 853 19.23   0.45
#> 854 19.23   1.45
#> 855 19.23   2.45
#> 856 19.23   3.45
#> 857 19.23   4.45
#> 858 19.23   5.45
#> 859 19.23   6.45
#> 860 19.23   7.45
#> 861 19.23   8.45
#> 862 19.23   9.45
#> 863 19.23  10.45
#> 864 19.23  11.45
#> 865 19.23  12.45
#> 866 19.23  13.45
#> 867 19.23  14.45
#> 868 19.23  15.45
#> 869 19.23  16.45
#> 870 19.23  17.45
#> 871 19.23  18.45
#> 872 19.23  19.45
#> 873 19.23  20.45
#> 874 19.23  21.45
#> 875 19.23  22.45

  head # => results in a data frame

#> function (x, ...) 
#> UseMethod("head")
#> <bytecode: 0x7f8d9604def8>
#> <environment: namespace:utils>

chondro %>%
  transmute (y, x, filename, spc = spc*2)

#> hyperSpec object
#>    875 spectra
#>    4 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 4 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. spc: I / a.u. [matrix, array300] 1003.639 1000.910 ... 338.5884 

  head # => results in a hyperSpec object

#> function (x, ...) 
#> UseMethod("head")
#> <bytecode: 0x7f8d9604def8>
#> <environment: namespace:utils>

flu %>%
  transmute (spc, filename, c)

#> hyperSpec object
#>    6 spectra
#>    3 data columns
#>    181 data points / spectrum
#> wavelength: lambda/nm [numeric] 405.0 405.5 ... 495 
#> data:  (6 rows x 3 columns)
#>    1. spc: I[fl]/"a.u." [matrix, array181] 27.15000 66.80133 ... 294.6495 
#>    2. filename: filename [character] rawdata/flu1.txt rawdata/flu2.txt ... rawdata/flu6.txt 
#>    3. c: c / (mg / l) [numeric] 0.05 0.10 ... 0.3 

  head # => results in a hyperSpec object

#> function (x, ...) 
#> UseMethod("head")
#> <bytecode: 0x7f8d9604def8>
#> <environment: namespace:utils>

flu %>%
  transmute (filename, c)

#>           filename    c
#> 1 rawdata/flu1.txt 0.05
#> 2 rawdata/flu2.txt 0.10
#> 3 rawdata/flu3.txt 0.15
#> 4 rawdata/flu4.txt 0.20
#> 5 rawdata/flu5.txt 0.25
#> 6 rawdata/flu6.txt 0.30

  head # => results in a data frame

#> function (x, ...) 
#> UseMethod("head")
#> <bytecode: 0x7f8d9604def8>
#> <environment: namespace:utils>

flu %>%
  transmute (c=c*2, c=c*0)

#>   c
#> 1 0
#> 2 0
#> 3 0
#> 4 0
#> 5 0
#> 6 0

  # => results in a data frame