transmute for hyperSpec object

transmute adds new variables and drops all pre-existing variables. Special column $spc contains the spectra matrix. If $spc is not being transmuted, the result is a data.frame instead of a hyperSpec object.

# S3 method for hyperSpec
transmute(.data, ...)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...
<data-masking> Name-value pairs. The name gives the name of the column in the output.

The value can be:

A vector of length 1, which will be recycled to the correct length.

A vector the same length as the current group (or the whole data frame if ungrouped).

NULL, to remove the column.

A data frame or tibble, to create multiple columns in the output.

Value

hyperSpec object.

See also

Examples

laser %>%
  transmute (t, filename)

#>       t             filename
#> 1     0 rawdata/laser.txt.gz
#> 2     2 rawdata/laser.txt.gz
#> 3    72 rawdata/laser.txt.gz
#> 4   142 rawdata/laser.txt.gz
#> 5   212 rawdata/laser.txt.gz
#> 6   281 rawdata/laser.txt.gz
#> 7   351 rawdata/laser.txt.gz
#> 8   421 rawdata/laser.txt.gz
#> 9   491 rawdata/laser.txt.gz
#> 10  560 rawdata/laser.txt.gz
#> 11  630 rawdata/laser.txt.gz
#> 12  699 rawdata/laser.txt.gz
#> 13  769 rawdata/laser.txt.gz
#> 14  839 rawdata/laser.txt.gz
#> 15  909 rawdata/laser.txt.gz
#> 16  979 rawdata/laser.txt.gz
#> 17 1049 rawdata/laser.txt.gz
#> 18 1119 rawdata/laser.txt.gz
#> 19 1189 rawdata/laser.txt.gz
#> 20 1259 rawdata/laser.txt.gz
#> 21 1328 rawdata/laser.txt.gz
#> 22 1399 rawdata/laser.txt.gz
#> 23 1468 rawdata/laser.txt.gz
#> 24 1538 rawdata/laser.txt.gz
#> 25 1608 rawdata/laser.txt.gz
#> 26 1678 rawdata/laser.txt.gz
#> 27 1748 rawdata/laser.txt.gz
#> 28 1818 rawdata/laser.txt.gz
#> 29 1888 rawdata/laser.txt.gz
#> 30 1957 rawdata/laser.txt.gz
#> 31 2026 rawdata/laser.txt.gz
#> 32 2096 rawdata/laser.txt.gz
#> 33 2166 rawdata/laser.txt.gz
#> 34 2236 rawdata/laser.txt.gz
#> 35 2305 rawdata/laser.txt.gz
#> 36 2375 rawdata/laser.txt.gz
#> 37 2445 rawdata/laser.txt.gz
#> 38 2515 rawdata/laser.txt.gz
#> 39 2584 rawdata/laser.txt.gz
#> 40 2654 rawdata/laser.txt.gz
#> 41 2724 rawdata/laser.txt.gz
#> 42 2793 rawdata/laser.txt.gz
#> 43 2863 rawdata/laser.txt.gz
#> 44 2933 rawdata/laser.txt.gz
#> 45 3003 rawdata/laser.txt.gz
#> 46 3073 rawdata/laser.txt.gz
#> 47 3143 rawdata/laser.txt.gz
#> 48 3212 rawdata/laser.txt.gz
#> 49 3281 rawdata/laser.txt.gz
#> 50 3351 rawdata/laser.txt.gz
#> 51 3421 rawdata/laser.txt.gz
#> 52 3491 rawdata/laser.txt.gz
#> 53 3561 rawdata/laser.txt.gz
#> 54 3630 rawdata/laser.txt.gz
#> 55 3700 rawdata/laser.txt.gz
#> 56 3770 rawdata/laser.txt.gz
#> 57 3840 rawdata/laser.txt.gz
#> 58 3909 rawdata/laser.txt.gz
#> 59 3979 rawdata/laser.txt.gz
#> 60 4049 rawdata/laser.txt.gz
#> 61 4119 rawdata/laser.txt.gz
#> 62 4189 rawdata/laser.txt.gz
#> 63 4259 rawdata/laser.txt.gz
#> 64 4328 rawdata/laser.txt.gz
#> 65 4398 rawdata/laser.txt.gz
#> 66 4467 rawdata/laser.txt.gz
#> 67 4537 rawdata/laser.txt.gz
#> 68 4607 rawdata/laser.txt.gz
#> 69 4677 rawdata/laser.txt.gz
#> 70 4747 rawdata/laser.txt.gz
#> 71 4817 rawdata/laser.txt.gz
#> 72 4886 rawdata/laser.txt.gz
#> 73 4956 rawdata/laser.txt.gz
#> 74 5026 rawdata/laser.txt.gz
#> 75 5095 rawdata/laser.txt.gz
#> 76 5164 rawdata/laser.txt.gz
#> 77 5234 rawdata/laser.txt.gz
#> 78 5304 rawdata/laser.txt.gz
#> 79 5373 rawdata/laser.txt.gz
#> 80 5443 rawdata/laser.txt.gz
#> 81 5513 rawdata/laser.txt.gz
#> 82 5583 rawdata/laser.txt.gz
#> 83 5652 rawdata/laser.txt.gz
#> 84 5722 rawdata/laser.txt.gz

  head # => results in a data frame

#> function (x, ...) 
#> UseMethod("head")
#> <bytecode: 0x7fa1b47d9cf8>
#> <environment: namespace:utils>

laser %>%
   transmute (-spc) # => results in a hyperSpec object

#>    -spc.-75.4628 -spc.-73.3067 -spc.-71.151 -spc.-68.9958 -spc.-66.841
#> 1      -164.6500     -190.1970    -286.5130     -307.4540    -310.9960
#> 2      -179.7240     -192.5170    -247.0740     -313.2550    -338.8470
#> 3      -183.2020     -215.7120    -288.8330     -337.6190    -327.2420
#> 4      -189.0000     -201.7950    -263.3140     -317.8960    -310.9960
#> 5      -187.8410     -200.6350    -263.3140     -321.3760    -351.6110
#> 6      -170.4480     -176.2810    -193.7150     -357.3420    -523.3560
#> 7      -160.0120     -187.8780    -226.1950     -396.7890    -527.9970
#> 8      -168.1290     -176.2810    -187.9150     -416.5130    -576.7360
#> 9      -149.5770     -177.4400    -244.7540     -390.9880    -516.3930
#> 10     -160.0120     -199.4750    -242.4340     -366.6240    -437.4840
#> 11     -157.6930     -197.1560    -257.5140     -323.6970    -398.0290
#> 12     -160.0120     -191.3570    -271.4330     -341.1000    -398.0290
#> 13     -183.2020     -167.0030    -287.6730     -316.7350    -367.8580
#> 14     -197.1170     -206.4340    -276.0730     -331.8180    -345.8090
#> 15     -177.4050     -215.7120    -313.1920     -335.2990    -337.6860
#> 16     -205.2330     -209.9130    -308.5530     -344.5800    -362.0550
#> 17     -170.4480     -202.9550    -281.8730     -352.7020    -351.6110
#> 18     -171.6070     -160.0440    -220.3950     -403.7510    -514.0720
#> 19     -165.8100     -154.2460    -245.9140     -416.5130    -574.4150
#> 20     -173.9260     -160.0440    -187.9150     -428.1150    -594.1420
#> 21     -158.8530     -177.4400    -193.7150     -461.7610    -645.2010
#> 22     -150.7360     -172.8010    -212.2750     -424.6340    -575.5750
#> 23     -169.2880     -190.1970    -221.5550     -394.4690    -495.5050
#> 24     -150.7360     -194.8360    -231.9940     -363.1430    -466.4940
#> 25     -170.4480     -212.2330    -236.6340     -374.7460    -398.0290
#> 26     -177.4050     -202.9550    -269.1130     -337.6190    -409.6330
#> 27     -190.1600     -194.8360    -302.7530     -345.7400    -409.6330
#> 28     -208.7120     -219.1910    -296.9530     -355.0220    -379.4620
#> 29     -195.9570     -204.1140    -306.2330     -341.1000    -371.3390
#> 30     -187.8410     -223.8300    -314.3520     -326.0170    -365.5370
#> 31     -193.6380     -221.5100    -301.5930     -357.3420    -357.4140
#> 32     -193.6380     -216.8720    -293.4730     -344.5800    -372.4990
#> 33     -185.5220     -204.1140    -262.1540     -327.1770    -399.1890
#> 34     -155.3740     -173.9610    -225.0350     -381.7070    -476.9380
#> 35     -162.3310     -167.0030    -191.3950     -428.1150    -576.7360
#> 36     -156.5340     -162.3640    -218.0750     -445.5180    -568.6130
#> 37     -153.0550     -190.1970    -220.3950     -481.4840    -618.5110
#> 38     -161.1720     -168.1620    -194.8750     -504.6880    -604.5860
#> 39     -147.2580     -160.0440    -211.1150     -483.8050    -638.2390
#> 40     -156.5340     -155.4050    -205.3150     -446.6780    -663.7680
#> 41     -139.1410     -164.6830    -205.3150     -459.4400    -613.8700
#> 42     -150.7360     -182.0790    -201.8350     -460.6010    -569.7730
#> 43     -166.9690     -169.3220    -212.2750     -430.4350    -537.2810
#> 44     -170.4480     -184.3990    -229.6740     -428.1150    -533.8000
#> 45     -182.0430     -172.8010    -222.7150     -394.4690    -522.1950
#> 46     -178.5640     -195.9960    -233.1540     -360.8230    -467.6550
#> 47     -180.8830     -189.0380    -218.0750     -377.0660    -447.9270
#> 48     -162.3310     -199.4750    -257.5140     -382.8670    -436.3230
#> 49     -198.2760     -185.5590    -254.0340     -350.3810    -409.6330
#> 50     -176.2450     -200.6350    -267.9530     -367.7840    -409.6330
#> 51     -195.9570     -193.6770    -278.3930     -365.4640    -416.5960
#> 52     -165.8100     -215.7120    -273.7530     -344.5800    -398.0290
#> 53     -163.4910     -220.3510    -273.7530     -365.4640    -401.5100
#> 54     -186.6810     -222.6700    -271.4330     -338.7790    -398.0290
#> 55     -183.2020     -213.3920    -296.9530     -350.3810    -380.6220
#> 56     -191.3190     -224.9900    -266.7940     -356.1820    -371.3390
#> 57     -198.2760     -208.7530    -278.3930     -343.4200    -386.4240
#> 58     -215.6690     -212.2330    -281.8730     -336.4590    -373.6600
#> 59     -168.1290     -231.9480    -279.5530     -353.8620    -369.0180
#> 60     -191.3190     -211.0730    -291.1530     -357.3420    -369.0180
#> 61     -193.6380     -211.0730    -296.9530     -351.5410    -378.3010
#> 62     -161.1720     -202.9550    -270.2730     -349.2210    -393.3870
#> 63     -193.6380     -212.2330    -262.1540     -365.4640    -385.2640
#> 64     -180.8830     -191.3570    -243.5940     -385.1870    -438.6440
#> 65     -168.1290     -191.3570    -247.0740     -396.7890    -445.6070
#> 66     -147.2580     -172.8010    -202.9950     -432.7560    -501.3070
#> 67     -161.1720     -160.0440    -204.1550     -406.0710    -552.3670
#> 68     -158.8530     -158.8840    -227.3540     -421.1540    -582.5380
#> 69     -163.4910     -144.9680    -214.5950     -490.7660    -588.3400
#> 70     -162.3310     -163.5230    -200.6750     -437.3960    -599.9440
#> 71     -150.7360     -164.6830    -193.7150     -468.7220    -586.0190
#> 72     -157.6930     -169.3220    -206.4750     -452.4790    -620.8320
#> 73     -158.8530     -157.7250    -213.4350     -484.9650    -604.5860
#> 74     -147.2580     -158.8840    -200.6750     -476.8430    -609.2280
#> 75     -149.5770     -164.6830    -206.4750     -444.3580    -620.8320
#> 76     -146.0980     -162.3640    -198.3550     -502.3680    -588.3400
#> 77     -157.6930     -165.8430    -202.9950     -469.8820    -615.0300
#> 78     -157.6930     -156.5650    -212.2750     -487.2850    -606.9070
#> 79     -161.1720     -143.8080    -186.7550     -473.3630    -577.8960
#> 80     -144.9390     -163.5230    -200.6750     -486.1250    -589.5000
#> 81     -153.0550     -156.5650    -190.2350     -442.0370    -575.5750
#> 82     -151.8960     -172.8010    -213.4350     -498.8870    -588.3400
#> 83     -155.3740     -180.9200    -182.1160     -457.1200    -561.6500
#> 84     -171.6070     -171.6420    -192.5550     -455.9600    -599.9440
#>    -spc.-64.6866 -spc.-62.5327 -spc.-60.3792 -spc.-58.2261 -spc.-56.0735
#> 1      -373.7340     -372.6470     -563.1460     -821.0790    -1118.6100
#> 2      -330.7890     -369.1650     -520.1850     -792.0460     -985.0260
#> 3      -371.4130     -384.2560     -543.4070     -790.8840    -1037.3000
#> 4      -358.6450     -381.9350     -530.6350     -775.7870    -1047.7500
#> 5      -383.0190     -364.5210     -618.8800     -860.5660     -998.9650
#> 6      -293.6480     -267.0060     -746.6040    -1296.0700     -880.4830
#> 7      -323.8250     -273.9710     -395.9440     -588.8080     -896.7460
#> 8      -311.0580     -264.6840     -315.8260     -443.6380     -777.1020
#> 9      -338.9140     -316.9250     -348.3380     -535.3860     -964.1180
#> 10     -351.6810     -439.9790     -499.2840     -661.9740    -1212.7000
#> 11     -371.4130     -436.4970     -512.0570     -763.0120    -1359.0600
#> 12     -364.4490     -441.1400     -531.7960     -781.5930    -1345.1200
#> 13     -376.0550     -442.3010     -573.5960     -841.9840    -1260.3200
#> 14     -381.8590     -407.4740     -617.7190     -897.7290    -1197.6000
#> 15     -389.9830     -390.0610     -607.2690     -915.1490    -1160.4300
#> 16     -377.2160     -394.7050     -610.7520     -877.9860    -1154.6200
#> 17     -366.7700     -363.3600     -623.5250     -945.3450    -1106.9900
#> 18     -304.0940     -258.8800     -775.6320    -1297.2400     -880.4830
#> 19     -331.9500     -276.2930     -418.0060     -646.8760     -885.1300
#> 20     -299.4520     -260.0410     -315.8260     -462.2200     -775.9410
#> 21     -293.6480     -236.8230     -297.2480     -434.3480     -725.9920
#> 22     -307.5760     -314.6030     -349.4990     -502.8680     -879.3220
#> 23     -354.0030     -459.7150     -477.2230     -653.8440    -1263.8100
#> 24     -380.6980     -473.6450     -516.7010     -743.2690    -1364.8700
#> 25     -377.2160     -472.4850     -537.6020     -835.0160    -1400.8700
#> 26     -384.1800     -466.6800     -588.6910     -868.6950    -1412.4900
#> 27     -373.7340     -430.6920     -604.9470     -895.4060    -1310.2700
#> 28     -387.6620     -402.8310     -630.4920     -909.3430    -1254.5100
#> 29     -421.3210     -428.3700     -644.4250     -963.9260    -1194.1100
#> 30     -406.2330     -393.5440     -632.8140     -965.0880    -1191.7900
#> 31     -427.1250     -427.2100     -663.0030     -987.1540    -1218.5100
#> 32     -396.9470     -405.1530     -645.5860     -963.9260    -1239.4100
#> 33     -371.4130     -369.1650     -707.1260    -1032.4500    -1188.3000
#> 34     -292.4880     -297.1890     -789.5660    -1198.5200     -953.6640
#> 35     -333.1110     -250.7530     -674.6140    -1013.8600     -844.4740
#> 36     -321.5040     -291.3850     -423.8110     -596.9380     -903.7150
#> 37     -315.7010     -270.4890     -350.6600     -504.0290     -828.2120
#> 38     -257.6680     -237.9840     -325.1150     -433.1860     -765.4860
#> 39     -297.1300     -268.1670     -297.2480     -421.5730     -752.7090
#> 40     -270.4350     -241.4660     -279.8310     -382.0860     -739.9310
#> 41     -293.6480     -261.2020     -294.9260     -465.7040     -800.3340
#> 42     -342.3960     -414.4400     -377.3660     -513.3200     -993.1570
#> 43     -350.5210     -467.8410     -411.0390     -580.6790    -1138.3600
#> 44     -316.8610     -477.1280     -444.7110     -613.1970    -1238.2500
#> 45     -364.4490     -477.1280     -491.1560     -697.9760    -1331.1800
#> 46     -370.2520     -469.0020     -505.0900     -740.9460    -1362.5400
#> 47     -349.3600     -473.6450     -543.4070     -792.0460    -1407.8400
#> 48     -388.8230     -464.3580     -557.3410     -843.1450    -1486.8300
#> 49     -387.6620     -481.7720     -579.4020     -841.9840    -1405.5200
#> 50     -388.8230     -446.9450     -646.7470     -874.5020    -1406.6800
#> 51     -362.1270     -482.9330     -623.5250     -882.6310    -1388.1000
#> 52     -408.5540     -456.2320     -654.8750     -915.1490    -1336.9900
#> 53     -432.9280     -422.5660     -611.9140     -897.7290    -1336.9900
#> 54     -381.8590     -445.7840     -656.0360     -943.0220    -1302.1400
#> 55     -407.3930     -431.8530     -659.5200     -941.8610    -1270.7800
#> 56     -376.0550     -414.4400     -680.4200     -980.1850    -1230.1200
#> 57     -386.5010     -394.7050     -674.6140    -1005.7400    -1238.2500
#> 58     -386.5010     -435.3360     -647.9090     -948.8290    -1283.5500
#> 59     -398.1080     -416.7620     -640.9420     -948.8290    -1205.7300
#> 60     -369.0910     -380.7740     -675.7760    -1006.9000    -1215.0200
#> 61     -394.6260     -399.3480     -674.6140     -972.0560    -1177.8500
#> 62     -400.4290     -372.6470     -675.7760    -1060.3200    -1248.7100
#> 63     -376.0550     -401.6700     -695.5150    -1033.6100    -1189.4700
#> 64     -352.8420     -351.7510     -695.5150    -1037.0900    -1073.3100
#> 65     -318.0220     -349.4300     -738.4760    -1097.4800     -989.6730
#> 66     -286.6840     -276.2930     -759.3770    -1110.2600     -882.8070
#> 67     -280.8810     -278.6150     -682.7420     -960.4420     -849.1210
#> 68     -299.4520     -263.5230     -591.0130     -902.3740     -854.9290
#> 69     -292.4880     -267.0060     -487.6730     -717.7190     -852.6050
#> 70     -271.5960     -254.2360     -392.4610     -596.9380     -840.9900
#> 71     -275.0780     -271.6500     -362.2710     -557.4510     -777.1020
#> 72     -305.2550     -251.9140     -335.5660     -512.1580     -829.3740
#> 73     -275.0780     -206.6390     -315.8260     -438.9930     -795.6880
#> 74     -297.1300     -263.5230     -300.7320     -451.7680     -751.5470
#> 75     -289.0060     -239.1440     -314.6650     -436.6700     -734.1230
#> 76     -285.5240     -241.4660     -287.9590     -408.7980     -735.2850
#> 77     -280.8810     -243.7880     -292.6040     -421.5730     -764.3250
#> 78     -285.5240     -242.6270     -277.5090     -408.7980     -693.4680
#> 79     -284.3630     -243.7880     -312.3430     -426.2180     -721.3460
#> 80     -292.4880     -232.1790     -277.5090     -408.7980     -727.1540
#> 81     -271.5960     -249.5930     -261.2530     -411.1200     -735.2850
#> 82     -304.0940     -250.7530     -290.2820     -413.4430     -736.4470
#> 83     -280.8810     -244.9490     -280.9930     -405.3140     -752.7090
#> 84     -266.9530     -255.3970     -291.4430     -401.8300     -727.1540
#>    -spc.-53.9212 -spc.-51.7694 -spc.-49.6181 -spc.-47.4671 -spc.-45.3166
#> 1     -2280.6500   -19159.9000   -56333.3000   -35176.4000    -9521.6800
#> 2     -2212.1000   -17674.8000   -51437.8000   -34295.2000    -9351.9200
#> 3     -2178.4100   -16291.9000   -47609.3000   -33911.6000    -9134.4800
#> 4     -2267.8700   -16153.6000   -47654.6000   -33202.4000    -9197.2700
#> 5     -2315.5000   -21480.5000   -58140.7000   -32389.8000    -8377.5400
#> 6     -2631.5200   -50316.7000   -73811.7000   -25682.2000    -2875.4600
#> 7     -2641.9700   -43381.6000   -73810.5000   -27866.5000    -4706.7700
#> 8     -2711.6800   -48937.4000   -73811.7000   -26213.4000    -3499.8500
#> 9     -2679.1500   -40511.3000   -73812.9000   -27830.5000    -5568.3600
#> 10    -2799.9800   -33762.2000   -73812.9000   -28947.6000    -8328.7000
#> 11    -2836.0000   -31167.3000   -73812.9000   -31391.2000   -10210.0000
#> 12    -2826.7000   -27635.8000   -69813.5000   -33067.6000   -10286.8000
#> 13    -2737.2400   -24978.2000   -64765.7000   -33574.4000    -9460.0500
#> 14    -2668.7000   -24361.2000   -62427.2000   -32737.4000    -8682.1800
#> 15    -2537.4100   -21989.4000   -57317.8000   -32066.7000    -8533.3500
#> 16    -2616.4100   -21260.8000   -55380.3000   -31286.6000    -8278.7100
#> 17    -2666.3700   -26757.3000   -63238.4000   -28612.8000    -7225.2600
#> 18    -3068.3600   -56609.2000   -73811.7000   -22460.8000    -2642.9100
#> 19    -3136.9100   -53302.0000   -73811.7000   -23844.2000    -3598.6800
#> 20    -3235.6600   -59528.3000   -73811.7000   -22283.0000    -2594.0700
#> 21    -3284.4600   -61646.7000   -73812.9000   -21208.8000    -2072.0000
#> 22    -3183.3800   -57362.2000   -73812.9000   -21948.2000    -3440.5500
#> 23    -3177.5700   -44627.3000   -73814.0000   -25650.8000    -8163.6000
#> 24    -3178.7300   -40222.0000   -73815.2000   -26514.5000    -9677.4800
#> 25    -3196.1600   -35987.5000   -73814.0000   -27867.7000    -9985.6100
#> 26    -3114.8300   -32162.0000   -71028.0000   -28433.8000    -9265.8700
#> 27    -3002.1400   -29778.7000   -68992.9000   -30014.8000    -8774.0300
#> 28    -2919.6500   -28738.6000   -66453.3000   -29312.7000    -7877.5600
#> 29    -2903.3800   -27009.5000   -62276.1000   -28910.4000    -7579.9000
#> 30    -2820.8900   -25142.1000   -58974.0000   -28259.4000    -7672.9200
#> 31    -2830.1900   -24952.7000   -58332.5000   -28667.5000    -7793.8400
#> 32    -2859.2300   -25421.0000   -58503.3000   -27479.4000    -7355.4900
#> 33    -2960.3100   -31882.0000   -68495.4000   -25722.8000    -6411.3500
#> 34    -3321.6400   -53764.5000   -73814.0000   -21363.4000    -3570.7700
#> 35    -3682.9600   -67712.6000   -73814.0000   -18690.8000    -1820.8500
#> 36    -3536.5700   -62489.2000   -73814.0000   -20164.9000    -2858.0100
#> 37    -3621.3900   -65741.7000   -73812.9000   -19369.7000    -2373.1500
#> 38    -3684.1300   -68732.9000   -73814.0000   -18217.7000    -1724.3400
#> 39    -3667.8600   -69796.1000   -73814.0000   -17744.5000    -1547.6100
#> 40    -3671.3500   -69897.2000   -73814.0000   -17922.4000    -1619.7000
#> 41    -3633.0100   -69970.4000   -73814.0000   -18447.9000    -2144.0900
#> 42    -3655.0800   -65738.3000   -73814.0000   -18592.0000    -4321.9000
#> 43    -3781.7200   -62660.0000   -73814.0000   -19727.8000    -6042.7600
#> 44    -3695.7400   -57916.5000   -73814.0000   -20538.1000    -6932.2500
#> 45    -3703.8800   -53529.8000   -73814.0000   -21422.7000    -8151.9700
#> 46    -3675.9900   -48335.4000   -73814.0000   -22288.8000    -8871.7000
#> 47    -3599.3100   -45596.5000   -73815.2000   -23123.5000    -9201.9200
#> 48    -3522.6300   -43424.6000   -73815.2000   -23618.7000    -9260.0600
#> 49    -3524.9600   -39904.8000   -73814.0000   -24447.6000    -9206.5700
#> 50    -3472.6700   -38137.3000   -73814.0000   -24869.6000    -8998.4400
#> 51    -3362.3000   -35743.5000   -70960.6000   -25091.6000    -8334.5200
#> 52    -3364.6300   -34222.3000   -68982.4000   -25042.8000    -8094.9900
#> 53    -3221.7200   -33920.2000   -68029.4000   -25245.1000    -7682.2200
#> 54    -3315.8300   -33683.1000   -68559.4000   -25146.2000    -7319.4500
#> 55    -3210.1000   -32609.4000   -66659.0000   -25028.8000    -6974.1100
#> 56    -3176.4100   -31956.3000   -64630.9000   -24706.8000    -6865.9800
#> 57    -3088.1100   -30910.5000   -62956.0000   -24870.7000    -6842.7200
#> 58    -3125.2900   -30326.0000   -62532.9000   -24767.3000    -6777.6100
#> 59    -3084.6300   -29344.1000   -60747.7000   -24432.5000    -6529.9500
#> 60    -3036.9900   -29708.9000   -60698.9000   -23900.0000    -6572.9700
#> 61    -3138.0700   -29593.9000   -60443.2000   -23678.0000    -6535.7600
#> 62    -3133.4200   -31481.1000   -63116.4000   -23283.9000    -6242.7500
#> 63    -3272.8400   -35092.7000   -67663.2000   -22582.9000    -5955.5500
#> 64    -3279.8100   -41008.7000   -73815.2000   -21272.8000    -4932.3400
#> 65    -3516.8200   -53723.8000   -73815.2000   -19022.1000    -3549.8400
#> 66    -3744.5400   -63838.3000   -73815.2000   -17909.6000    -2354.5500
#> 67    -3863.0500   -69561.4000   -73815.2000   -16973.8000    -1685.9700
#> 68    -4018.7300   -71534.6000   -73816.3000   -17199.3000    -1777.8300
#> 69    -4052.4200   -71277.7000   -73816.3000   -17105.2000    -1869.6900
#> 70    -4073.3300   -72418.9000   -73816.3000   -16733.2000    -1744.1100
#> 71    -3986.2000   -71950.6000   -73816.3000   -16629.7000    -1816.2000
#> 72    -3985.0400   -71896.0000   -73816.3000   -16607.6000    -1642.9500
#> 73    -4033.8300   -73073.1000   -73816.3000   -16523.9000    -1487.1400
#> 74    -3979.2300   -73421.7000   -73816.3000   -16365.8000    -1481.3300
#> 75    -3960.6400   -72983.6000   -73817.5000   -15898.5000    -1369.7100
#> 76    -4023.3800   -73801.7000   -73816.3000   -16041.5000    -1418.5400
#> 77    -3876.9900   -72530.4000   -73817.5000   -15892.7000    -1362.7300
#> 78    -3967.6100   -72695.4000   -73817.5000   -15726.4000    -1318.5500
#> 79    -3936.2400   -73801.7000   -73816.3000   -15897.3000    -1372.0300
#> 80    -3936.2400   -73428.7000   -73817.5000   -15795.0000    -1346.4500
#> 81    -3967.6100   -73069.6000   -73816.3000   -15521.8000    -1361.5700
#> 82    -3836.3200   -72531.6000   -73817.5000   -15566.0000    -1462.7300
#> 83    -3843.2900   -72517.7000   -73817.5000   -15750.8000    -1381.3300
#> 84    -3814.2500   -73219.5000   -73816.3000   -15348.6000    -1429.0100
#>    -spc.-43.1665 -spc.-41.0168 -spc.-38.8676 -spc.-36.7188 -spc.-34.5704
#> 1    -10664.4000   -21114.4000   -11516.8000    -5536.6900   -14807.0000
#> 2    -10670.2000   -21661.1000   -12106.7000    -5616.9900   -14009.7000
#> 3    -10968.0000   -22775.5000   -12942.0000    -5969.5800   -13865.4000
#> 4    -11248.2000   -23108.1000   -13131.6000    -5905.5700   -14056.3000
#> 5    -10778.4000   -22316.0000   -12193.9000    -5569.2800   -15930.2000
#> 6     -5395.0200   -11676.2000    -4232.5600    -2064.3300   -22331.6000
#> 7     -4636.7600    -8368.0600    -5266.8500    -3146.5400   -21505.2000
#> 8     -3055.1200    -5004.0900    -3485.6400    -2475.1000   -24533.6000
#> 9     -5243.8300    -9150.8900    -5900.9200    -3523.5600   -22167.4000
#> 10    -8222.2000   -13416.3000    -8047.4500    -4334.6300   -19213.5000
#> 11   -10053.9000   -16310.4000    -8952.5900    -4860.6100   -18808.5000
#> 12   -10748.2000   -18808.9000    -9855.4200    -5231.8200   -18112.5000
#> 13   -10872.6000   -20678.2000   -10900.2000    -5373.7800   -17374.5000
#> 14   -11237.8000   -22412.5000   -11480.7000    -5550.6600   -17662.0000
#> 15   -11790.2000   -24267.8000   -12346.3000    -5928.8500   -17213.9000
#> 16   -12185.6000   -24894.8000   -12801.2000    -6059.1800   -17429.3000
#> 17   -11194.7000   -22739.4000   -11031.6000    -5400.5500   -18631.5000
#> 18    -5764.8400   -11441.2000    -3756.7200    -2135.3200   -26151.4000
#> 19    -4280.8900    -7632.9200    -4116.2200    -2788.1300   -26802.1000
#> 20    -2248.0200    -3474.4800    -2215.1700    -2205.1400   -29848.0000
#> 21    -1403.7000    -1957.6700    -1484.5400    -1930.5100   -31412.2000
#> 22    -2931.8500    -4278.2500    -2765.4700    -2546.0900   -30531.2000
#> 23    -7455.8000   -10922.4000    -6101.0300    -3963.4300   -24250.8000
#> 24    -9329.3400   -13944.4000    -7260.9700    -4457.9800   -23013.6000
#> 25   -10431.8000   -16401.1000    -8306.8900    -4925.7700   -22217.5000
#> 26   -10985.4000   -18749.6000    -9357.4700    -5051.4500   -21463.3000
#> 27   -11721.6000   -21372.6000   -10472.0000    -5422.6600   -20723.1000
#> 28   -11703.0000   -22278.8000   -10614.0000    -5565.7900   -20240.0000
#> 29   -12304.2000   -23537.4000   -11201.5000    -5605.3500   -20156.2000
#> 30   -12989.2000   -24832.0000   -11851.9000    -5990.5200   -19975.8000
#> 31   -13489.3000   -26082.4000   -12200.9000    -6073.1400   -19847.8000
#> 32   -13153.2000   -25417.1000   -11919.3000    -5926.5200   -20153.9000
#> 33   -12296.1000   -23608.3000   -10327.8000    -5290.0000   -22202.4000
#> 34    -7972.1600   -14986.7000    -5397.1500    -3126.7500   -28280.2000
#> 35    -4034.3400    -7102.5000    -2123.2600    -1904.9100   -32341.0000
#> 36    -3585.4400    -5869.5100    -2887.6300    -2513.5100   -31795.1000
#> 37    -2348.0400    -3482.6200    -2002.2600    -2315.6800   -33904.1000
#> 38    -1260.6600    -1527.2800    -1170.4100    -1943.3100   -35823.4000
#> 39     -980.3840     -999.1890     -900.4950    -1923.5300   -37345.7000
#> 40     -915.2570     -837.5040     -846.9770    -1946.8000   -37774.0000
#> 41    -1394.4000    -1422.5900    -1149.4700    -2053.8600   -37387.6000
#> 42    -2816.7100    -2581.1400    -1873.1200    -2318.0100   -34680.4000
#> 43    -4302.9900    -4480.6500    -2943.4800    -2746.2400   -32882.2000
#> 44    -5522.9400    -6666.3000    -3880.0400    -3174.4600   -31255.1000
#> 45    -7151.1000    -8928.7200    -4794.5000    -3544.5100   -29431.3000
#> 46    -8595.5100   -11466.8000    -5804.3500    -3941.3200   -27528.3000
#> 47    -9379.3500   -13322.1000    -6653.6600    -4300.8900   -26948.7000
#> 48   -10302.8000   -14905.2000    -7220.2500    -4554.5700   -26043.2000
#> 49   -10802.8000   -16444.2000    -8017.2000    -4764.0200   -25035.3000
#> 50   -11235.5000   -18023.8000    -8437.1900    -5080.5400   -24509.2000
#> 51   -11437.8000   -19075.3000    -8988.6600    -5192.2500   -23424.4000
#> 52   -11677.4000   -20309.5000    -9483.1200    -5288.8300   -23187.0000
#> 53   -12151.9000   -21105.1000    -9671.5900    -5469.2000   -23445.4000
#> 54   -12486.8000   -22185.7000   -10075.3000    -5533.2000   -23530.4000
#> 55   -12391.4000   -22555.6000   -10261.5000    -5355.1600   -22879.7000
#> 56   -12639.2000   -22872.0000   -10439.5000    -5528.5500   -22754.0000
#> 57   -12988.0000   -24003.8000   -10733.8000    -5648.4100   -22812.2000
#> 58   -13383.5000   -24549.3000   -10990.9000    -5803.1700   -22811.1000
#> 59   -13219.5000   -24620.3000   -10965.3000    -5833.4300   -22354.8000
#> 60   -13348.6000   -24702.9000   -11057.2000    -5720.5500   -22539.9000
#> 61   -13425.3000   -24768.0000   -11139.8000    -5767.1000   -22477.0000
#> 62   -13316.0000   -24480.7000   -10710.5000    -5591.3900   -23682.8000
#> 63   -12942.7000   -23822.3000    -9837.9600    -5465.7100   -24742.0000
#> 64   -11509.9000   -20858.5000    -8422.0700    -4537.1100   -26541.3000
#> 65    -8666.4500   -15379.8000    -5618.2000    -3460.7200   -30779.1000
#> 66    -6554.5000   -11176.0000    -3498.4300    -2575.1800   -33929.7000
#> 67    -4356.4900    -7112.9700    -2101.1500    -2188.8400   -36169.0000
#> 68    -3466.8100    -5562.4200    -1820.7700    -2142.3000   -36756.8000
#> 69    -2774.8500    -4270.1100    -1719.5500    -2248.1900   -36585.7000
#> 70    -2286.4000    -3180.1900    -1461.2700    -2272.6300   -37014.0000
#> 71    -1981.7000    -2689.3200    -1350.7400    -2185.3500   -37556.4000
#> 72    -1630.4800    -2160.0600    -1191.3500    -2115.5300   -37754.3000
#> 73    -1182.7400    -1359.7800     -945.8690    -2062.0100   -38783.1000
#> 74    -1045.5100    -1096.9000     -877.2260    -2056.1900   -39416.3000
#> 75     -921.0720     -850.2990     -793.4590    -2093.4200   -39443.1000
#> 76     -881.5310     -856.1150     -801.6030    -2110.8800   -40307.8000
#> 77     -768.7230     -737.4690     -759.7200    -2041.0600   -40228.7000
#> 78     -788.4940     -643.2500     -744.5950    -2089.9300   -41121.4000
#> 79     -796.6340     -668.8400     -736.4510    -2135.3200   -41348.4000
#> 80     -808.2640     -687.4510     -743.4320    -2112.0400   -41417.0000
#> 81     -785.0050     -668.8400     -736.4510    -2119.0200   -42053.7000
#> 82     -797.7970     -649.0660     -798.1130    -2132.9900   -42008.3000
#> 83     -826.8720     -647.9020     -732.9610    -2122.5200   -40909.6000
#> 84     -862.9240     -673.4930     -757.3930    -2193.5000   -41763.9000
#>    -spc.-32.4224 -spc.-30.2749 -spc.-28.1277 -spc.-25.981 -spc.-23.8348
#> 1    -34901.5000   -17290.6000   -24019.4000  -15054.0000    -5987.1300
#> 2    -32560.4000   -17249.8000   -24444.5000  -15420.9000    -6142.0800
#> 3    -31623.3000   -17857.6000   -26621.1000  -16430.8000    -6777.0300
#> 4    -31367.2000   -18182.5000   -27363.0000  -16607.8000    -7089.2600
#> 5    -37169.2000   -18333.8000   -24835.8000  -15013.2000    -6342.4700
#> 6    -60552.9000   -14840.8000    -5262.7300   -3226.5200    -1831.4400
#> 7    -57038.4000   -16659.5000   -11129.9000   -6511.2800    -2667.9400
#> 8    -66106.9000   -16731.7000    -7460.2900   -4427.4400    -1783.6700
#> 9    -57371.3000   -17189.3000   -12960.6000   -7643.4800    -3019.7800
#> 10   -47016.5000   -16781.8000   -18672.9000  -10969.0000    -4020.5500
#> 11   -44529.9000   -17046.1000   -20853.0000  -11981.2000    -4438.8000
#> 12   -41668.5000   -17375.6000   -22318.0000  -13133.2000    -4996.8500
#> 13   -39386.8000   -17426.8000   -23773.7000  -13945.1000    -5727.3300
#> 14   -38907.2000   -17917.0000   -24523.7000  -14277.1000    -6085.0000
#> 15   -36955.0000   -18495.7000   -26753.9000  -15280.0000    -6894.7000
#> 16   -36871.2000   -19349.2000   -28244.5000  -16268.9000    -7288.4800
#> 17   -41130.7000   -18487.5000   -23809.8000  -13676.0000    -6105.9700
#> 18   -65721.5000   -14401.8000    -5051.9400   -2987.7300    -1798.8200
#> 19   -65090.6000   -15701.3000    -8912.5200   -5004.0200    -2112.2100
#> 20   -73763.3000   -15393.9000    -5035.6400   -2744.2900    -1186.0100
#> 21   -73929.8000   -15269.3000    -3101.2700   -1688.9700     -702.5180
#> 22   -73929.8000   -15289.1000    -6832.5900   -3722.7300    -1483.0900
#> 23   -56258.4000   -15938.8000   -15650.8000   -8547.3700    -3126.9600
#> 24   -52886.0000   -16247.3000   -18431.8000  -10168.8000    -3573.1700
#> 25   -49571.7000   -16663.0000   -20516.4000  -11277.7000    -4391.0300
#> 26   -45406.5000   -17051.9000   -22214.3000  -12284.1000    -5136.6500
#> 27   -43172.5000   -17749.3000   -23847.1000  -13188.0000    -5807.7200
#> 28   -42511.3000   -18000.8000   -24866.1000  -13504.8000    -6152.5700
#> 29   -41130.7000   -18473.6000   -26317.2000  -14203.7000    -6602.2700
#> 30   -39354.2000   -19145.4000   -27451.5000  -14989.9000    -6938.9700
#> 31   -39539.3000   -19676.3000   -28514.7000  -15366.2000    -7295.4700
#> 32   -40108.6000   -19370.1000   -27735.6000  -14970.1000    -7092.7500
#> 33   -45420.5000   -18694.8000   -23555.9000  -12776.8000    -6173.5400
#> 34   -62609.9000   -14897.9000   -10015.4000   -5565.4600    -3022.1100
#> 35   -73140.5000   -12873.0000    -2687.8500   -1538.7100    -1066.0100
#> 36   -69735.4000   -14339.0000    -6397.0300   -3403.5700    -1525.0400
#> 37   -73929.8000   -14156.2000    -4481.3000   -2409.9900    -1000.7700
#> 38   -73930.9000   -14094.5000    -2534.1200   -1280.1200     -539.4130
#> 39   -73930.9000   -14169.0000    -1933.2000    -978.4390     -442.7150
#> 40   -73930.9000   -14250.5000    -1789.9600    -906.2210     -380.9680
#> 41   -73930.9000   -14386.7000    -2682.0300   -1391.9500     -505.6270
#> 42   -73930.9000   -14144.5000    -5126.4800   -2564.9100     -850.4780
#> 43   -71032.3000   -14421.6000    -8177.6700   -4234.0800    -1437.6600
#> 44   -67585.3000   -14696.4000   -10902.8000   -5628.3600    -1880.3700
#> 45   -63072.0000   -15345.0000   -13459.0000   -7040.1000    -2445.4200
#> 46   -58199.0000   -15688.4000   -15764.9000   -8252.6700    -2920.7500
#> 47   -55530.8000   -16163.5000   -17517.6000   -9253.2400    -3465.9900
#> 48   -53117.6000   -16674.6000   -18781.2000   -9892.7200    -3813.1700
#> 49   -50400.6000   -16866.8000   -20340.5000  -10658.0000    -4386.3700
#> 50   -48825.5000   -17608.5000   -21443.4000  -11388.3000    -4758.0200
#> 51   -46605.5000   -17521.1000   -22065.3000  -11914.8000    -5203.0600
#> 52   -45772.0000   -17873.9000   -23087.8000  -12363.3000    -5469.8600
#> 53   -45158.5000   -18133.6000   -23590.9000  -12633.5000    -5875.2900
#> 54   -45467.0000   -18438.6000   -24082.3000  -12842.0000    -5958.0100
#> 55   -44767.4000   -18550.4000   -23964.7000  -12893.3000    -6053.5400
#> 56   -44089.9000   -18790.3000   -24541.2000  -13130.9000    -6255.0900
#> 57   -43720.8000   -18916.0000   -25533.4000  -13553.7000    -6463.6300
#> 58   -43474.1000   -19541.3000   -26348.6000  -13807.6000    -6825.9600
#> 59   -42830.3000   -19558.7000   -25817.6000  -13934.6000    -6759.5500
#> 60   -42755.8000   -19557.6000   -26067.9000  -13709.8000    -6765.3800
#> 61   -42709.2000   -19630.9000   -26078.4000  -13874.0000    -6800.3300
#> 62   -44514.8000   -19498.2000   -24890.6000  -13195.0000    -6510.2300
#> 63   -47973.4000   -18959.1000   -23249.7000  -12258.4000    -6072.1800
#> 64   -52159.5000   -17812.2000   -18938.4000  -10051.1000    -5063.2600
#> 65   -62451.5000   -15501.0000   -11246.3000   -6069.8200    -3281.9100
#> 66   -69669.1000   -13901.2000    -5817.0700   -3254.4800    -1972.4100
#> 67   -73932.1000   -13203.7000    -2744.9100   -1531.7200    -1088.1500
#> 68   -73933.2000   -13441.3000    -2838.0800   -1525.9000     -985.6220
#> 69   -73933.2000   -13535.6000    -3128.0600   -1697.1300     -904.0700
#> 70   -73933.2000   -13365.6000    -2984.8200   -1558.5100     -749.1200
#> 71   -73933.2000   -13493.6000    -2819.4500   -1426.8900     -665.2370
#> 72   -73933.2000   -13534.4000    -2518.9800   -1343.0200     -566.2090
#> 73   -73933.2000   -13707.9000    -1986.7700   -1001.7400     -445.0450
#> 74   -73933.2000   -13508.8000    -1762.0100    -889.9140     -383.2980
#> 75   -73934.4000   -13601.9000    -1678.1600    -827.0140     -363.4920
#> 76   -73933.2000   -13606.6000    -1646.7100    -845.6510     -368.1520
#> 77   -73934.4000   -13748.6000    -1559.3700    -760.6200     -383.2980
#> 78   -73934.4000   -13809.2000    -1513.9500    -729.1700     -351.8420
#> 79   -73933.2000   -14001.3000    -1511.6200    -729.1700     -321.5510
#> 80   -73934.4000   -14046.7000    -1511.6200    -736.1590     -361.1620
#> 81   -73933.2000   -13806.9000    -1543.0700    -730.3350     -337.8610
#> 82   -73934.4000   -14088.6000    -1597.8000    -766.4440     -371.6470
#> 83   -73934.4000   -13501.8000    -1569.8500    -752.4660     -348.3460
#> 84   -73933.2000   -13308.5000    -1522.1000    -745.4780     -329.7060
#>    -spc.-21.6889 -spc.-19.5435 -spc.-17.3985 -spc.-15.2539 -spc.-13.1097
#> 1     -5940.5400    -2783.2100    -1443.1700    -1188.1100    -1173.1800
#> 2     -6287.7900    -2920.7400    -1490.9600    -1252.2400    -1221.0000
#> 3     -6838.9600    -3186.4700    -1622.6900    -1337.3500    -1307.2900
#> 4     -6976.4600    -3315.8400    -1680.9800    -1460.9400    -1356.2700
#> 5     -6323.9100    -2991.8300    -1513.1100    -1258.0700    -1197.6700
#> 6     -1406.4800     -709.7880     -417.3300     -408.0850     -367.3480
#> 7     -2512.3200    -1314.6800     -748.3960     -654.1010     -586.5910
#> 8     -1613.9000     -904.4260     -538.5660     -538.6720     -427.9900
#> 9     -2835.1000    -1498.8300     -799.6880     -756.7050     -618.0780
#> 10    -3802.2700    -1850.8100    -1011.8500     -844.1520     -784.8430
#> 11    -4100.5800    -1979.0100    -1112.1000     -908.2800     -882.8020
#> 12    -4729.8200    -2327.5000    -1201.8600     -986.3990     -957.4380
#> 13    -5437.1400    -2567.5900    -1346.4100    -1143.8000    -1075.2200
#> 14    -5905.5800    -2664.3300    -1438.5100    -1235.9100    -1155.6900
#> 15    -6663.0000    -3017.4700    -1564.4100    -1340.8500    -1278.1400
#> 16    -6981.1200    -3175.9800    -1693.8000    -1404.9800    -1393.5900
#> 17    -5811.1900    -2589.7300    -1380.2200    -1186.9400    -1146.3600
#> 18    -1352.8800     -608.3900     -425.4900     -462.8840     -366.1820
#> 19    -1961.1500     -970.8590     -587.5260     -582.9780     -514.2880
#> 20    -1014.9500     -579.2520     -413.8330     -424.4080     -340.5260
#> 21     -658.3760     -412.5860     -331.0660     -373.1060     -254.2280
#> 22    -1342.3900     -724.9390     -488.4390     -507.1910     -362.6840
#> 23    -2851.4100    -1442.8900     -823.0030     -735.7180     -663.5590
#> 24    -3369.9500    -1607.2200     -880.1240     -811.5050     -712.5390
#> 25    -3998.0300    -1907.9200    -1047.9900     -907.1140     -817.4960
#> 26    -4715.8400    -2160.8300    -1186.7100    -1041.2000     -935.2800
#> 27    -5440.6400    -2441.7200    -1306.7800    -1169.4500    -1128.8700
#> 28    -5793.7100    -2595.5600    -1419.8600    -1247.5700    -1146.3600
#> 29    -6249.3300    -2867.1200    -1479.3100    -1326.8600    -1264.1400
#> 30    -6700.2900    -2951.0400    -1640.1800    -1443.4500    -1302.6300
#> 31    -7041.7100    -3055.9300    -1655.3300    -1452.7800    -1369.1000
#> 32    -6593.0900    -2906.7500    -1604.0400    -1430.6300    -1362.1000
#> 33    -5611.9300    -2550.1100    -1369.7300    -1161.2900    -1160.3500
#> 34    -2416.7700    -1109.5500     -699.4360     -664.5950     -551.6060
#> 35     -657.2110     -339.1600     -293.7630     -321.8040     -243.7330
#> 36    -1296.9400     -694.6360     -474.4510     -495.5310     -389.5060
#> 37     -871.6200     -511.6530     -367.2040     -401.0890     -297.3770
#> 38     -467.2720     -315.8500     -289.1000     -361.4460     -254.2280
#> 39     -348.4150     -248.2510     -263.4540     -341.6250     -201.7500
#> 40     -315.7880     -247.0850     -243.6370     -370.7740     -215.7440
#> 41     -467.2720     -317.0150     -286.7690     -360.2800     -239.0680
#> 42     -709.6480     -412.5860     -341.5580     -368.4420     -309.0390
#> 43    -1215.3700     -648.0160     -425.4900     -458.2210     -346.3570
#> 44    -1728.0900     -855.4750     -549.0570     -556.1610     -418.6600
#> 45    -2148.7500    -1040.7900     -621.3320     -571.3180     -522.4510
#> 46    -2595.0500    -1238.9200     -755.3910     -712.3990     -611.0810
#> 47    -3019.2100    -1440.5500     -847.4830     -743.8800     -653.0640
#> 48    -3404.9100    -1571.0900     -920.9240     -855.8120     -710.2070
#> 49    -3880.3400    -1803.0200    -1031.6700     -925.7690     -838.4870
#> 50    -4337.1300    -2022.1400    -1108.6100     -949.0880     -911.9570
#> 51    -4823.0400    -2146.8500    -1197.2000    -1059.8500     -993.5900
#> 52    -5159.8100    -2285.5400    -1271.8100    -1150.8000    -1064.7300
#> 53    -5370.7200    -2416.0800    -1372.0600    -1155.4600    -1086.8800
#> 54    -5693.5000    -2472.0200    -1354.5700    -1228.9200    -1110.2100
#> 55    -5695.8300    -2481.3400    -1387.2100    -1233.5800    -1180.1800
#> 56    -5868.2900    -2559.4300    -1428.0200    -1287.2200    -1211.6700
#> 57    -6019.7700    -2672.4900    -1401.2000    -1332.6900    -1248.9800
#> 58    -6305.2600    -2759.9000    -1541.0900    -1400.3100    -1276.9700
#> 59    -6265.6500    -2744.7500    -1497.9600    -1368.8300    -1232.6600
#> 60    -6163.1000    -2680.6400    -1523.6000    -1307.0400    -1268.8100
#> 61    -6203.8900    -2705.1200    -1479.3100    -1340.8500    -1286.3000
#> 62    -5875.2800    -2599.0600    -1403.5300    -1331.5200    -1189.5100
#> 63    -5397.5200    -2425.4000    -1338.2500    -1256.9000    -1107.8800
#> 64    -4398.8900    -1963.8600    -1102.7800    -1033.0400     -921.2860
#> 65    -2551.9400    -1164.3300     -715.7560     -676.2540     -607.5820
#> 66    -1407.6400     -656.1750     -477.9480     -481.5400     -376.6780
#> 67     -619.9220     -365.9660     -308.9180     -352.1190     -239.0680
#> 68     -618.7570     -371.7940     -319.4090     -341.6250     -249.5640
#> 69     -659.5420     -372.9590     -339.2260     -376.6040     -244.8990
#> 70     -570.9810     -372.9590     -318.2430     -356.7830     -235.5690
#> 71     -522.0400     -328.6700     -301.9230     -355.6170     -230.9050
#> 72     -418.3310     -314.6840     -277.4430     -362.6120     -239.0680
#> 73     -346.0850     -257.5750     -287.9340     -355.6170     -228.5720
#> 74     -308.7960     -234.2650     -285.6030     -317.1400     -215.7440
#> 75     -273.8380     -255.2440     -266.9510     -331.1310     -201.7500
#> 76     -288.9860     -250.5820     -263.4540     -339.2930     -211.0800
#> 77     -255.1940     -226.1060     -257.6260     -326.4680     -195.9190
#> 78     -278.4990     -216.7820     -258.7910     -342.7910     -213.4120
#> 79     -273.8380     -216.7820     -259.9570     -325.3020     -207.5810
#> 80     -257.5240     -250.5820     -278.6090     -317.1400     -199.4180
#> 81     -281.9950     -208.6240     -251.7970     -329.9660     -207.5810
#> 82     -285.4910     -228.4370     -262.2880     -314.8080     -198.2510
#> 83     -241.2100     -233.0990     -266.9510     -346.2890     -206.4150
#> 84     -248.2020     -224.9410     -262.2880     -319.4720     -209.9130
#>    -spc.-10.966 -spc.-8.82266 -spc.-6.67976 -spc.-4.53728 -spc.-2.39522
#> 1     -618.2000     -410.6580     -220.5390     -154.0570     -140.0800
#> 2     -600.7030     -404.8250     -260.2130     -170.3970     -127.2390
#> 3     -690.5170     -412.9920     -274.2150     -193.7390     -126.0720
#> 4     -695.1830     -438.6580     -274.2150     -179.7340     -134.2430
#> 5     -627.5310     -408.3250     -239.2090     -193.7390     -147.0840
#> 6     -240.2810     -186.6630     -145.8590     -131.8820     -113.2310
#> 7     -355.7560     -230.9950     -176.1980     -134.2170     -117.9000
#> 8     -257.7780     -194.8290     -143.5250     -140.0520     -103.8920
#> 9     -384.9170     -258.9950     -170.3630     -134.2170     -128.4060
#> 10    -485.2280     -306.8270     -200.7020     -147.0550     -119.0680
#> 11    -489.8940     -319.6600     -215.8710     -159.8930     -136.5780
#> 12    -577.3750     -332.4930     -235.7080     -159.8930     -152.9200
#> 13    -607.7020     -365.1590     -248.5440     -152.8900     -137.7450
#> 14    -635.6960     -400.1590     -275.3820     -161.0600     -135.4100
#> 15    -709.1800     -417.6580     -264.8800     -177.3990     -148.2510
#> 16    -724.3430     -437.4910     -282.3830     -179.7340     -161.0920
#> 17    -575.0420     -368.6590     -259.0460     -175.0650     -141.2470
#> 18    -230.9500     -163.3300     -142.3580     -124.8800     -115.5660
#> 19    -297.4360     -202.9960     -180.8650     -150.5560     -123.7370
#> 20    -232.1160     -174.9970     -157.5280     -135.3840     -116.7330
#> 21    -184.2930     -159.8300     -129.5230     -131.8820     -128.4060
#> 22    -249.6130     -205.3290     -142.3580     -130.7150     -121.4020
#> 23    -400.0800     -250.8280     -197.2010     -152.8900     -126.0720
#> 24    -416.4100     -290.4940     -194.8680     -142.3860     -131.9080
#> 25    -456.0680     -338.3270     -197.2010     -159.8930     -129.5740
#> 26    -545.8820     -356.9930     -225.2060     -157.5590     -130.7410
#> 27    -624.0320     -405.9920     -250.8780     -187.9030     -151.7530
#> 28    -603.0360     -383.8260     -239.2090     -177.3990     -151.7530
#> 29    -648.5260     -402.4920     -270.7140     -179.7340     -130.7410
#> 30    -664.8560     -438.6580     -280.0490     -168.0630     -149.4180
#> 31    -692.8500     -453.8240     -256.7120     -184.4020     -150.5860
#> 32    -660.1900     -426.9910     -269.5480     -170.3970     -140.0800
#> 33    -590.2060     -382.6590     -275.3820     -171.5640     -157.5900
#> 34    -341.7590     -232.1620     -175.0310     -163.3940     -126.0720
#> 35    -199.4570     -160.9970     -129.5230     -113.2090     -117.9000
#> 36    -262.4430     -191.3300     -148.1930     -137.7180     -116.7330
#> 37    -215.7870     -179.6630     -136.5240     -128.3810     -113.2310
#> 38    -176.1290     -181.9960     -135.3570     -134.2170     -130.7410
#> 39    -170.2960     -134.1640     -141.1920     -117.8770     -124.9040
#> 40    -176.1290     -142.3300     -135.3570     -127.2140     -122.5700
#> 41    -183.1270     -169.1630     -138.8580     -142.3860     -120.2350
#> 42    -188.9590     -171.4970     -134.1900     -129.5480      -99.2231
#> 43    -253.1120     -193.6630     -134.1900     -120.2110     -131.9080
#> 44    -278.7730     -248.4950     -164.5290     -124.8800     -121.4020
#> 45    -367.4200     -223.9960     -176.1980     -158.7260     -138.9120
#> 46    -386.0830     -265.9950     -171.5300     -169.2300     -136.5780
#> 47    -408.2450     -288.1610     -173.8640     -135.3840     -128.4060
#> 48    -439.7380     -303.3270     -217.0380     -158.7260     -147.0840
#> 49    -481.7290     -311.4940     -218.2050     -168.0630     -143.5820
#> 50    -514.3890     -330.1600     -231.0410     -152.8900     -148.2510
#> 51    -516.7220     -333.6600     -231.0410     -154.0570     -148.2510
#> 52    -573.8760     -363.9930     -238.0420     -161.0600     -163.4260
#> 53    -598.3710     -362.8260     -255.5450     -155.2240     -147.0840
#> 54    -621.6990     -375.6590     -250.8780     -178.5670     -149.4180
#> 55    -635.6960     -362.8260     -249.7110     -168.0630     -137.7450
#> 56    -619.3660     -391.9920     -255.5450     -176.2320     -155.2550
#> 57    -668.3550     -407.1590     -275.3820     -170.3970     -119.0680
#> 58    -632.1970     -398.9920     -276.5490     -184.4020     -148.2510
#> 59    -640.3610     -417.6580     -255.5450     -178.5670     -162.2590
#> 60    -632.1970     -416.4920     -257.8790     -185.5690     -136.5780
#> 61    -641.5280     -407.1590     -239.2090     -187.9030     -148.2510
#> 62    -631.0300     -395.4920     -241.5430     -189.0700     -151.7530
#> 63    -599.5370     -365.1590     -255.5450     -161.0600     -135.4100
#> 64    -507.3900     -348.8260     -212.3710     -175.0650     -143.5820
#> 65    -347.5910     -261.3280     -199.5350     -149.3890     -123.7370
#> 66    -251.9450     -178.4960     -142.3580     -143.5530     -136.5780
#> 67    -191.2920     -160.9970     -147.0260     -136.5510     -106.2270
#> 68    -192.4580     -155.1640     -130.6900     -122.5460     -127.2390
#> 69    -184.2930     -165.6630     -131.8570     -131.8820     -123.7370
#> 70    -188.9590     -186.6630     -138.8580     -129.5480     -122.5700
#> 71    -159.7990     -174.9970     -129.5230     -138.8850     -110.8960
#> 72    -188.9590     -146.9970     -133.0230     -138.8850     -116.7330
#> 73    -172.6290     -150.4970     -140.0250     -136.5510     -129.5740
#> 74    -184.2930     -138.8310     -142.3580     -129.5480     -124.9040
#> 75    -163.2980     -159.8300     -130.6900     -115.5430     -117.9000
#> 76    -171.4630     -144.6640     -141.1920     -124.8800     -137.7450
#> 77    -151.6340     -151.6640     -117.8540     -129.5480     -120.2350
#> 78    -153.9670     -145.8300     -140.0250     -120.2110     -114.3980
#> 79    -171.4630     -162.1630     -135.3570      -98.0365     -123.7370
#> 80    -171.4630     -141.1640     -136.5240     -128.3810     -117.9000
#> 81    -146.9680     -149.3300     -136.5240     -137.7180     -122.5700
#> 82    -162.1320     -127.1640     -130.6900     -122.5460     -122.5700
#> 83    -153.9670     -151.6640     -129.5230     -124.8800     -141.2470
#> 84    -173.7960     -141.1640     -138.8580     -119.0440      -98.0557
#>    -spc.-0.253579
#> 1       -130.7670
#> 2       -124.9290
#> 3       -123.7610
#> 4       -123.7610
#> 5       -136.6040
#> 6       -110.9180
#> 7       -120.2590
#> 8       -131.9340
#> 9       -106.2480
#> 10      -114.4210
#> 11      -115.5880
#> 12      -116.7560
#> 13      -120.2590
#> 14      -124.9290
#> 15      -128.4320
#> 16      -138.9400
#> 17      -129.5990
#> 18      -103.9130
#> 19      -119.0910
#> 20      -110.9180
#> 21      -120.2590
#> 22      -114.4210
#> 23      -110.9180
#> 24      -133.1020
#> 25      -127.2640
#> 26      -129.5990
#> 27      -131.9340
#> 28      -150.6150
#> 29      -129.5990
#> 30      -145.9450
#> 31      -149.4480
#> 32      -129.5990
#> 33      -142.4420
#> 34      -135.4370
#> 35      -102.7450
#> 36      -117.9230
#> 37       -99.2425
#> 38       -98.0750
#> 39      -102.7450
#> 40      -120.2590
#> 41      -141.2750
#> 42      -121.4260
#> 43      -108.5830
#> 44      -122.5940
#> 45      -113.2530
#> 46      -110.9180
#> 47      -122.5940
#> 48      -134.2690
#> 49      -115.5880
#> 50      -141.2750
#> 51      -143.6100
#> 52      -123.7610
#> 53      -121.4260
#> 54      -129.5990
#> 55      -148.2800
#> 56      -128.4320
#> 57      -142.4420
#> 58      -124.9290
#> 59      -133.1020
#> 60      -131.9340
#> 61      -147.1120
#> 62      -140.1070
#> 63      -141.2750
#> 64      -135.4370
#> 65      -136.6040
#> 66      -123.7610
#> 67      -128.4320
#> 68      -126.0960
#> 69      -108.5830
#> 70      -130.7670
#> 71      -108.5830
#> 72      -121.4260
#> 73      -114.4210
#> 74      -116.7560
#> 75      -101.5780
#> 76      -100.4100
#> 77      -119.0910
#> 78      -129.5990
#> 79      -119.0910
#> 80      -102.7450
#> 81      -115.5880
#> 82      -119.0910
#> 83      -115.5880
#> 84      -112.0860

laser %>%
   transmute (spc2 = spc*2) %>%
   transmute (spc2) %>%
   transmute (spc2*2) # => results in a hyperSpec object

#> hyperSpec object
#>    84 spectra
#>    1 data columns
#>    36 data points / spectrum
#> wavelength: lambda/nm [numeric] 404.5828 404.6181 ... 405.8176 
#> data:  (84 rows x 1 columns)
#>    1. spc2 * 2:  [matrix, array36] 658.600 718.896 ... 448.344